Create Draft 2
Instead of saving the info to a notepad I decided to have it save to an excel for easier viewing and editing. Next I will make it so that when a new posting is made under 500$ it will send me an email. and have the program run once a day to update me